fix from copilot comment
